Moving GzipSttpBackend
